third_party/blink/web_tests/platform/linux-chrome/external/wpt/dom/nodes/Document-createEvent.https-expected.txt

This is a testharness.js-based test.
Found 16 FAIL, 0 TIMEOUT, 0 NOTRUN.
[FAIL] Should throw NOT_SUPPORTED_ERR for pluralized legacy event interface "KeyboardEvents"
  assert_throws_dom: function "function () {\n          var evt = document.createEvent(plural);\n        }" did not throw
[FAIL] TextEvent should be an alias for CompositionEvent.
  assert_equals: expected object "[object CompositionEvent]" but got object "[object TextEvent]"
[FAIL] textevent should be an alias for CompositionEvent.
  assert_equals: expected object "[object CompositionEvent]" but got object "[object TextEvent]"
[FAIL] TEXTEVENT should be an alias for CompositionEvent.
  assert_equals: expected object "[object CompositionEvent]" but got object "[object TextEvent]"
[PRECONDITION_FAILED] TouchEvent should be an alias for TouchEvent.
  'expose legacy touch event APIs'
[PRECONDITION_FAILED] createEvent('TouchEvent') should be initialized correctly.
  'expose legacy touch event APIs'
[PRECONDITION_FAILED] touchevent should be an alias for TouchEvent.
  'expose legacy touch event APIs'
[PRECONDITION_FAILED] createEvent('touchevent') should be initialized correctly.
  'expose legacy touch event APIs'
[PRECONDITION_FAILED] TOUCHEVENT should be an alias for TouchEvent.
  'expose legacy touch event APIs'
[PRECONDITION_FAILED] createEvent('TOUCHEVENT') should be initialized correctly.
  'expose legacy touch event APIs'
[FAIL] Should throw NOT_SUPPORTED_ERR for non-legacy event interface "AnimationEvent"
  assert_throws_dom: function "function () {\n      var evt = document.createEvent(eventInterface);\n    }" did not throw
[FAIL] Should throw NOT_SUPPORTED_ERR for non-legacy event interface "CloseEvent"
  assert_throws_dom: function "function () {\n      var evt = document.createEvent(eventInterface);\n    }" did not throw
[FAIL] Should throw NOT_SUPPORTED_ERR for non-legacy event interface "ErrorEvent"
  assert_throws_dom: function "function () {\n      var evt = document.createEvent(eventInterface);\n    }" did not throw
[FAIL] Should throw NOT_SUPPORTED_ERR for non-legacy event interface "IDBVersionChangeEvent"
  assert_throws_dom: function "function () {\n      var evt = document.createEvent(eventInterface);\n    }" did not throw
[FAIL] Should throw NOT_SUPPORTED_ERR for non-legacy event interface "MutationEvent"
  assert_throws_dom: function "function () {\n      var evt = document.createEvent(eventInterface);\n    }" did not throw
[FAIL] Should throw NOT_SUPPORTED_ERR for pluralized non-legacy event interface "MutationEvents"
  assert_throws_dom: function "function () {\n        var evt = document.createEvent(eventInterface + "s");\n      }" did not throw
[FAIL] Should throw NOT_SUPPORTED_ERR for non-legacy event interface "PageTransitionEvent"
  assert_throws_dom: function "function () {\n      var evt = document.createEvent(eventInterface);\n    }" did not throw
[FAIL] Should throw NOT_SUPPORTED_ERR for non-legacy event interface "PopStateEvent"
  assert_throws_dom: function "function () {\n      var evt = document.createEvent(eventInterface);\n    }" did not throw
[FAIL] Should throw NOT_SUPPORTED_ERR for non-legacy event interface "TrackEvent"
  assert_throws_dom: function "function () {\n      var evt = document.createEvent(eventInterface);\n    }" did not throw
[FAIL] Should throw NOT_SUPPORTED_ERR for non-legacy event interface "TransitionEvent"
  assert_throws_dom: function "function () {\n      var evt = document.createEvent(eventInterface);\n    }" did not throw
[FAIL] Should throw NOT_SUPPORTED_ERR for non-legacy event interface "WebGLContextEvent"
  assert_throws_dom: function "function () {\n      var evt = document.createEvent(eventInterface);\n    }" did not throw
[FAIL] Should throw NOT_SUPPORTED_ERR for non-legacy event interface "WheelEvent"
  assert_throws_dom: function "function () {\n      var evt = document.createEvent(eventInterface);\n    }" did not throw
Harness: the test ran to completion.
